Python/requirements.txt
flask
gunicorn
mysql-connector-python
numpy
opencv-python-headless
face_recognition
